trybot_control: Use the right pull request number in the trybot payload.
We should pass the ID of the PullRequest database entry, not the GitHub pull
request number: there may be multiple rows with the same GitHub pull request
number. The views code was also already assuming that the "issue" property
we send back and forth to Buildbot refers to a database ID, and would fail
if there was no database entry with the same number as a GitHub pull
request.

Fix that by introducing some code that is maybe too fragile: we partially
construct a payload in make_trybot_payload(), create a new PullRequest and
then use its primary key as the 'number' entry in the Buildbot payload.