trybot_control: Report build status immediately.
Instead of waiting for the first build to start, set a pull request's status
to pending immediately when it is created. This makes more sense, as it
might take a while for the patch to be processed and we can consider that
the pull request is pending from the beginning.