jira_updater: Stop assuming there is only one ID for the Resolve transition.
It turns out that the ID of the "Resolve" transition varies according to the
current state an issue is at (it may not even exist).

Take that into account and now query the server for the ID corresponding to
that transition. It introduces some latency as it means another query to the
JIRA server, but it should be bearable for now.

BUG=XWALK-1078