trybot_control: Skip processing if the patch is not for master.
The Buildbot infrastructure cannot handle patches for different branches at
the moment, and would always end up trying to apply them to master in the
slaves.

For now, work around the issue here as we'd otherwise needlessly add a
comment to the pull requests saying they would be tested, and perform some
useless network I/O to fetch patches and things like that.

In the future, we should move this kind of detection to Buildbot itself.

BUG=XWALK-1077