jira_updater: Fix behavior when multiple issues are mentioned.
This commit addresses two different issues in search_issues() that are more
easily fixed together:

1. If multiple issues were mentioned in the same line, only the first one
   would be updated because we were calling re.search() once for each line
   and thus grabbing only the first match.

2. If the same issue was first mentioned in a normal paragraph and then in a
   BUG= line it would not be closed because the first mention would set
   'resolve' to False.

We now use re.findall() to go through all matches in a line to address 1),
and update the value of 'resolve' in the issues dict to fix 2).

A bunch of unit tests has also been added as well.

BUG=XWALK-1549
BUG=XWALK-1550