jira_updater: Make the template strings unicode ones.
open_comment_template and close_comment_template were non-unicode
strings, so we were effectively calling str.format() instead of
unicode.format() in methods such as comment_issue() or resolve_issue().

This is usually not a problem since all the text in the pull requests is
usually in English, but it breaks with Crosswalk pull request #1909, for
example, which uses non-ASCII unicode characters for the quotes in the
title.

Fix the issue by making {open,close}_comment_template proper unicode
strings.