trybot_control: Link to the user's fork when mentioning a hash.
Right now, when one submits a pull request and the hooks code writes a
comment saying "Testing patch series 1234abcd" or "Patch series 1234abcd
will be tested soon", it uses the hash in the target repository (such as
crosswalk-project/crosswalk) instead of the source one (such as
someuser/crosswalk-fork).

Linking to the latter allows one to see comments made to commits
themselves more easily.

This introduces a change to the PullRequest model that requires
migrating existing databases.

BUG=XWALK-1890