trybot_control: Switch to fetching .diff files from GitHub.
The diff files provided by GitHub contain the diff between the target
branch and the source one in the pull request instead of containing the
commits from the source branch like the .patch files we used to use do.

This distinction is important because sometimes the original commits
fail to apply while the diff works as expected. For example, in
crosswalk-project/chromium-crosswalk#216 we are trying to merge a merge
commit which had to solve some conflicts. The diff applies correctly,
while the .patch file does not because it contains only the original
commits (which cause conflicts when using `git am').