trybot_control: Process patches to crosswalk's crosswalk-lite branch.
This is a temporary hack introduced to allow builders to process patches
targeted at crosswalk's crosswalk-lite branch.

In the future, it will be replaced by a proper workflow where we only
post comments and store a PullRequest object if there are builders to
process a patch targeting a certain branch.

BUG=XWALK-3347