jira_updater: Fix test_resolve_issue().
We were mistakenly calling assertTrue() instead of assertEqual() to
compare two values, so the test was broken.

Consequently, also update the expected values that we are comparing, as
some of those were not taking into consideration the previous call to
self.client.post(). Add some assertEqual() calls in that place too to
show what we expect to happen.