trybot_control: Properly quote builder name in the status URL
Something changed in GitHub a few weeks ago and it no longer seems to be
escaping URLs on its own (e.g. "foo bar" does not become "foo%20bar"). This
broke the tables we post when reporting a builder's status when the
builder's name contains spaces.

Make sure we quote the name correctly before posting it.